mywal/requirements.txt
joblib==1.5.1
numpy==2.3.2
pillow==11.3.0
scikit-learn==1.7.1
scipy==1.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
